requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
cloudinary==1.41.0
distlib==0.3.9
dj-database-url==2.2.0
Django==5.1.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3
django-rest-passwordreset==1.4.1
djangorestframework==3.15.2
filelock==3.16.1
idna==3.10
Markdown==3.7
pillow==10.4.0
platformdirs==4.3.6
psycopg2-binary==2.9.9
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
sslcommerz-lib==1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.28.0
whitenoise==6.7.0
